+++ tests/testthat/test-driver-rds.R
@@ -381,3 +381,11 @@ test_that("avoid race condition when writing in parallel", {
   ok <- vlapply(1:10, function(i) racy_write())
   expect_true(all(ok))
 })
+
+test_that("use_scratch = FALSE (#116)", {
+  st <- storr_rds(tempfile(), use_scratch = FALSE)
+  st$set("a", "a")
+  expect_equal(st$get("a"), "a")
+  expect_null(st$driver$path_scratch)
+  expect_false(st$driver$use_scratch)
+})