+++ NEWS.md
@@ -1,6 +1,7 @@
 ## storr 1.2.2 (2018-??-??)
 
 * Speed up the `$get_hash()` method of RDS drivers using C code and traits (#96, #98, @wlandau).
+* Allow users to bypass scratch in RDS `storr`s (#116, @wlandau).
 
 ## storr 1.2.1 (2018-10-18)